pyproject.toml
[project]
name = "news_aggregator"
version = "0.1.0"
description = "An agent-driven news monitoring system for financial, geopolitical, and macroeconomic insights."
authors = [{ name = "Yun Ye Jin" }]
dependencies = [
    "python-telegram-bot[job-queue]>=20.0",
    "APScheduler>=3.10.0",
    "feedparser>=6.0.10",
    "aiohttp>=3.9.0",
    "aiosqlite>=0.19.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "PyYAML>=6.0.1",
    "openai>=1.0.0",
    "sentence-transformers>=2.2.2",
    "chromadb>=0.4.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
